requirements.txt
discord.py==2.5.2
python-dotenv==1.1.1
requests==2.32.4
fastapi[standard]==0.115.14
websockets==15.0.1
aiohttp==3.12.13
uvicorn==0.35.0